refactor(endpoints): unify tab structure and consolidate components
