+++ apps/frontend/src/app/(protected)/endpoints/[identifier]/components/EndpointDetailTabs.tsx
@@ -12,16 +12,14 @@ import {
 } from './endpoint-detail-shared';
 import EndpointOverviewTab from './EndpointOverviewTab';
 import EndpointConnectionTab from './EndpointConnectionTab';
-import EndpointHeadersTab from './EndpointHeadersTab';
 import EndpointMappingTab from './EndpointMappingTab';
 import EndpointTestTab from './EndpointTestTab';
 
 const TAB_LABELS: Record<EndpointTabKey, string> = {
   overview: 'Overview',
   connection: 'Connection',
-  headers: 'Headers',
   mapping: 'Mapping',
-  'connection-test': 'Connection Test',
+  test: 'Test',
 };
 
 function normalizeTabParam(param: string | null): EndpointTabKey {
@@ -76,12 +74,9 @@ export default function EndpointDetailTabs() {
         <EndpointConnectionTab />
       </DetailTabPanel>
       <DetailTabPanel value={activeTab} index={2} prefix="endpoint-detail">
-        <EndpointHeadersTab />
-      </DetailTabPanel>
-      <DetailTabPanel value={activeTab} index={3} prefix="endpoint-detail">
         <EndpointMappingTab />
       </DetailTabPanel>
-      <DetailTabPanel value={activeTab} index={4} prefix="endpoint-detail">
+      <DetailTabPanel value={activeTab} index={3} prefix="endpoint-detail">
         <EndpointTestTab />
       </DetailTabPanel>
     </Box>
