+++ apps/frontend/src/app/(protected)/endpoints/[identifier]/components/endpoint-detail-shared.tsx
@@ -7,18 +7,13 @@ export {
 
 export const METHODS = ['POST'] as const;
 
-export const TAB_KEYS = [
-  'overview',
-  'connection',
-  'headers',
-  'mapping',
-  'connection-test',
-] as const;
+export const TAB_KEYS = ['overview', 'connection', 'mapping', 'test'] as const;
 export type EndpointTabKey = (typeof TAB_KEYS)[number];
 
-/** @deprecated Old tab URLs — `mappings` was the former combined test tab */
 export const LEGACY_TAB_MAP: Record<string, EndpointTabKey> = {
-  mappings: 'connection-test',
+  mappings: 'test',
+  'connection-test': 'test',
+  headers: 'connection',
 };
 
 export function getEnvironmentChipColor():
