+++ apps/frontend/src/app/(protected)/endpoints/components/EndpointForm.tsx
@@ -24,10 +24,10 @@ import { useNotifications } from '@/components/common/NotificationContext';
 import { useOnboarding } from '@/contexts/OnboardingContext';
 import { readActiveProjectId } from '@/utils/active-project';
 import AutoConfigureDrawer from './AutoConfigureDrawer';
-import TabBasics from './tabs/TabBasics';
-import TabHeaders from './tabs/TabHeaders';
-import TabBody from './tabs/TabBody';
-import TabTest from './tabs/TabTest';
+import TabOverview from './TabOverview';
+import TabConnection from './TabConnection';
+import TabMapping from './TabMapping';
+import TabTest from './TabTest';
 import {
   bodyToRequestMapping,
   parseBodyMapping,
@@ -58,9 +58,9 @@ const DEFAULT_RES_BODY = '{}';
 
 const CREATE_TABS = [
   { key: 'overview', label: 'Overview' },
-  { key: 'headers', label: 'Headers' },
+  { key: 'connection', label: 'Connection' },
   { key: 'mapping', label: 'Mapping' },
-  { key: 'connection-test', label: 'Connection Test' },
+  { key: 'test', label: 'Test' },
 ] as const;
 
 function validateUrl(url: string) {
@@ -128,7 +128,6 @@ const EndpointForm = forwardRef<EndpointFormHandle, EndpointFormProps>(
     const [error, setError] = useState<string | null>(null);
     const [projects, setProjects] = useState<Project[]>([]);
     const [loadingProjects, setLoadingProjects] = useState(true);
-    const [showAuthToken, setShowAuthToken] = useState(false);
     const [isSubmitting, setIsSubmitting] = useState(false);
     const [autoConfigureOpen, setAutoConfigureOpen] = useState(false);
 
@@ -451,7 +450,7 @@ const EndpointForm = forwardRef<EndpointFormHandle, EndpointFormProps>(
 
         <Box sx={{ flex: 1, minHeight: 0, overflow: 'auto' }}>
           <DetailTabPanel value={activeTab} index={0} prefix="endpoint">
-            <TabBasics
+            <TabOverview
               formData={formData}
               onChange={handleChange}
               projects={projects}
@@ -461,16 +460,11 @@ const EndpointForm = forwardRef<EndpointFormHandle, EndpointFormProps>(
           </DetailTabPanel>
 
           <DetailTabPanel value={activeTab} index={1} prefix="endpoint">
-            <TabHeaders
-              formData={formData}
-              onChange={handleChange}
-              showAuthToken={showAuthToken}
-              onToggleAuthToken={() => setShowAuthToken(v => !v)}
-            />
+            <TabConnection formData={formData} onChange={handleChange} />
           </DetailTabPanel>
 
           <DetailTabPanel value={activeTab} index={2} prefix="endpoint">
-            <TabBody
+            <TabMapping
               reqBody={reqBody}
               resBody={resBody}
               onReqBodyChange={setReqBody}
