+++ apps/frontend/src/app/(protected)/endpoints/components/__tests__/TabBasics.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import TabBasics from '../tabs/TabBasics';
+import TabOverview from '../TabOverview';
 import type { FormData } from '../EndpointForm';
 import type { Project } from '@/utils/api-client/interfaces/project';
 
@@ -27,22 +27,16 @@ const projects = [
   { id: 'proj-2', name: 'Beta', description: 'Second' },
 ] as unknown as Project[];
 
-describe('TabBasics', () => {
+describe('TabOverview', () => {
   it('renders required text fields', () => {
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={jest.fn()}
         projects={projects}
         loadingProjects={false}
       />
     );
-    expect(screen.getByRole('textbox', { name: /method/i })).toHaveValue(
-      'POST'
-    );
-    expect(
-      screen.getByRole('textbox', { name: /endpoint url/i })
-    ).toBeInTheDocument();
     expect(
       screen.getByRole('textbox', { name: /endpoint name/i })
     ).toBeInTheDocument();
@@ -53,7 +47,7 @@ describe('TabBasics', () => {
 
   it('hides the project dropdown when hideProjectSelect is true', () => {
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={jest.fn()}
         projects={projects}
@@ -69,7 +63,7 @@ describe('TabBasics', () => {
 
   it('shows a warning when hideProjectSelect is true but no project is resolved', () => {
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={jest.fn()}
         projects={projects}
@@ -85,7 +79,7 @@ describe('TabBasics', () => {
 
   it('disables the project dropdown while projects are loading', () => {
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={jest.fn()}
         projects={[]}
@@ -100,7 +94,7 @@ describe('TabBasics', () => {
 
   it('shows no-projects warning when list is empty and not loading', () => {
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={jest.fn()}
         projects={[]}
@@ -113,34 +107,10 @@ describe('TabBasics', () => {
     ).toBeInTheDocument();
   });
 
-  it('shows URL validation error for malformed URL', () => {
-    render(
-      <TabBasics
-        formData={{ ...defaultFormData, url: 'not-a-url' }}
-        onChange={jest.fn()}
-        projects={projects}
-        loadingProjects={false}
-      />
-    );
-    expect(screen.getByText(/enter a valid url/i)).toBeInTheDocument();
-  });
-
-  it('does not show URL error for valid URL', () => {
-    render(
-      <TabBasics
-        formData={{ ...defaultFormData, url: 'https://api.example.com' }}
-        onChange={jest.fn()}
-        projects={projects}
-        loadingProjects={false}
-      />
-    );
-    expect(screen.queryByText(/enter a valid url/i)).not.toBeInTheDocument();
-  });
-
   it('calls onChange with name when endpoint name field changes', () => {
     const onChange = jest.fn();
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={onChange}
         projects={projects}
@@ -153,26 +123,10 @@ describe('TabBasics', () => {
     expect(onChange).toHaveBeenCalledWith('name', 'My API');
   });
 
-  it('calls onChange with url when URL field changes', () => {
-    const onChange = jest.fn();
-    render(
-      <TabBasics
-        formData={defaultFormData}
-        onChange={onChange}
-        projects={projects}
-        loadingProjects={false}
-      />
-    );
-    fireEvent.change(screen.getByRole('textbox', { name: /endpoint url/i }), {
-      target: { value: 'https://api.example.com' },
-    });
-    expect(onChange).toHaveBeenCalledWith('url', 'https://api.example.com');
-  });
-
   it('calls onChange with description when description field changes', () => {
     const onChange = jest.fn();
     render(
-      <TabBasics
+      <TabOverview
         formData={defaultFormData}
         onChange={onChange}
         projects={projects}
