+++ apps/frontend/src/app/(protected)/endpoints/components/__tests__/TabBody.test.tsx
@@ -3,14 +3,14 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import userEvent from '@testing-library/user-event';
 
-jest.mock('../TestAndMap', () => {
+jest.mock('../MappingEditor', () => {
   return {
     __esModule: true,
     default: () => <div data-testid="test-and-map" />,
   };
 });
 
-import TabBody from '../tabs/TabBody';
+import TabMapping from '../TabMapping';
 
 const defaultProps = {
   reqBody: '{"input": "{{ input }}"}',
@@ -23,9 +23,9 @@ const defaultProps = {
   onAutoConfigureOpen: jest.fn(),
 };
 
-describe('TabBody', () => {
+describe('TabMapping', () => {
   it('renders the Auto Mapping button', () => {
-    render(<TabBody {...defaultProps} />);
+    render(<TabMapping {...defaultProps} />);
     expect(
       screen.getByRole('button', { name: /auto mapping/i })
     ).toBeInTheDocument();
@@ -35,19 +35,19 @@ describe('TabBody', () => {
     const onAutoConfigureOpen = jest.fn();
     const user = userEvent.setup({ delay: null });
     render(
-      <TabBody {...defaultProps} onAutoConfigureOpen={onAutoConfigureOpen} />
+      <TabMapping {...defaultProps} onAutoConfigureOpen={onAutoConfigureOpen} />
     );
     await user.click(screen.getByRole('button', { name: /auto mapping/i }));
     expect(onAutoConfigureOpen).toHaveBeenCalledTimes(1);
   });
 
   it('renders the Manual Mapping section', () => {
-    render(<TabBody {...defaultProps} />);
+    render(<TabMapping {...defaultProps} />);
     expect(screen.getByText(/manual mapping/i)).toBeInTheDocument();
   });
 
   it('expands Manual Mapping section when expand control is clicked', async () => {
-    render(<TabBody {...defaultProps} />);
+    render(<TabMapping {...defaultProps} />);
     fireEvent.click(
       screen.getByRole('button', { name: /expand manual mapping/i })
     );
@@ -55,7 +55,7 @@ describe('TabBody', () => {
   });
 
   it('renders the Auto Mapping description text', () => {
-    render(<TabBody {...defaultProps} />);
+    render(<TabMapping {...defaultProps} />);
     expect(
       screen.getByText(/paste your api docs or a sample response/i)
     ).toBeInTheDocument();
